+++ src/config/seatnotifications.config.php
@@ -24,5 +24,5 @@
  */
 
 return [
-    'version'   => '1.1.1',
+    'version'   => '1.1.2',
 ];